Make IAS client as separate module
- This helps reusage of ias client by both SGX worker and Graphene worker
